backend/pyproject.toml
[project]
name = "ai-cskh-backend"
version = "0.1.0"
description = "Customer service agent backend built with FastAPI and LangGraph"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.135.2",
  "uvicorn[standard]>=0.42.0",
  "langgraph>=1.1.3",
  "langchain>=1.2.13",
  "langchain-openai>=1.1.12",
  "llama-index>=0.14.19",
  "llama-index-vector-stores-faiss>=0.6.0",
  "faiss-cpu>=1.13.2",
  "sqlalchemy>=2.0.48",
  "pydantic-settings>=2.13.1",
  "httpx>=0.28.1",
  "tenacity>=9.1.4",
  "orjson>=3.11.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.15.7",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["app*"]
